more decimal

+++ db/migrate/20160810134633_create_bookings.rb
@@ -3,7 +3,7 @@ class CreateBookings < ActiveRecord::Migration[5.0]
     create_table :bookings do |t|
       t.datetime :begin_date
       t.datetime :end_date
-      t.float :total_price
+      t.decimal :total_price
 
       t.timestamps
     end
